+++ src/test/java/org/peterjaycruz/realworld/api/constants/Path.java
