+++ src/test/java/org/peterjaycruz/realworld/api/utilities/StringUtilities.java
